Update throttle and regen encoding This commit uses the full [0, 255] range for the throttle digital POT. It fixes an earlier bug where only a part of the input range was used.

+++ Motor/lib/MotorInterface/MotorInterface.h
@@ -8,8 +8,8 @@ class MotorInterface {
   public:
     MotorInterface(I2C &throttle, I2C &regen, DigitalOut &gear,
                    DigitalOut &ignition);
-    int sendThrottle(uint8_t throttle);
-    int sendRegen(uint8_t regen);
+    int sendThrottle(uint16_t throttle);
+    int sendRegen(uint16_t regen);
     void sendDirection(bool direction);
     void sendIgnition(bool ignition);
